refactor: move react components into react directory
