+++ packages/frontend/src/components/Header/HomeContent.astro
@@ -6,7 +6,7 @@
 // replaced on navigation, so a `document` lookup is enough for the target).
 import { SITE_NAME } from '@@frontend/constants/site';
 
-import ScrollButton from './ScrollButton';
+import ScrollButton from '../react/ScrollButton';
 ---
 
 <div class="layout">
