+++ packages/frontend/src/components/Header/index.astro
@@ -9,7 +9,7 @@
 // scroll hint is a plain Astro component with a small `<script>` (the whole
 // header is replaced on navigation, so a `document` lookup is enough for the
 // scroll target).
-import Nav from '@@frontend/components/Nav';
+import Nav from '@@frontend/components/react/Nav';
 
 import Content from './Content.astro';
 import HeaderBackground from './HeaderBackground.astro';
