+++ packages/frontend/src/components/react/FlexSpace/FlexSpace.module.scss
