+++ packages/frontend/src/layouts/PostLayout.astro
@@ -6,7 +6,7 @@ import Author from '@@frontend/components/Author.astro';
 import H1 from '@@frontend/components/H1.astro';
 import Main from '@@frontend/components/Main.astro';
 import MainContainer from '@@frontend/components/MainContainer.astro';
-import PostComments from '@@frontend/components/PostComments';
+import PostComments from '@@frontend/components/react/PostComments';
 import BaseLayout from '@@frontend/layouts/BaseLayout.astro';
 
 interface Props {
