+++ packages/frontend/src/layouts/RootLayout.astro
@@ -5,7 +5,7 @@
 import Box from '@@frontend/components/Box.astro';
 import Footer from '@@frontend/components/Footer.astro';
 import Header from '@@frontend/components/Header/index.astro';
-import SyncTheme from '@@frontend/components/SyncTheme';
+import SyncTheme from '@@frontend/components/react/SyncTheme';
 
 interface Props {
   headerKind: 'home' | 'default';
